v0.3.0

+++ backend/src/main/java/com/github/viktor235/rggassistant/models/dto/GameOnPlatformDto.java
@@ -1,14 +1,14 @@
 package com.github.viktor235.rggassistant.models.dto;
 
-import com.github.viktor235.rggassistant.models.Game;
-import com.github.viktor235.rggassistant.models.Platform;
+import com.github.viktor235.rggassistant.models.entitys.games.Game;
+import com.github.viktor235.rggassistant.models.entitys.games.Platform;
 import lombok.Builder;
 import lombok.Data;
 
 @Data
 @Builder
 public class GameOnPlatformDto {
-    private int id;
+    private long id;
     private Game game;
     private Platform platform;
 }
