+++ backend/src/main/java/com/github/viktor235/rggassistant/repositories/CollectedEffectRepository.java
@@ -1,9 +1,9 @@
 package com.github.viktor235.rggassistant.repositories;
 
-import com.github.viktor235.rggassistant.models.CollectedEffect;
+import com.github.viktor235.rggassistant.models.entitys.modifiers.CollectedEffect;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CollectedEffectRepository extends JpaRepository<CollectedEffect, Integer> {
+public interface CollectedEffectRepository extends JpaRepository<CollectedEffect, Long> {
 }
