+++ backend/src/main/java/com/github/viktor235/rggassistant/repositories/CollectedGameRepository.java
@@ -1,9 +1,9 @@
 package com.github.viktor235.rggassistant.repositories;
 
-import com.github.viktor235.rggassistant.models.CollectedGamePlatform;
+import com.github.viktor235.rggassistant.models.entitys.games.CollectedGamePlatform;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CollectedGameRepository extends JpaRepository<CollectedGamePlatform, Integer> {
+public interface CollectedGameRepository extends JpaRepository<CollectedGamePlatform, Long> {
 }
