+++ backend/src/main/java/com/github/viktor235/rggassistant/repositories/GameRepository.java
@@ -1,6 +1,6 @@
 package com.github.viktor235.rggassistant.repositories;
 
-import com.github.viktor235.rggassistant.models.Game;
+import com.github.viktor235.rggassistant.models.entitys.games.Game;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface GameRepository extends JpaRepository<Game, Integer> {
+public interface GameRepository extends JpaRepository<Game, Long> {
     @Query("SELECT g FROM Game g ORDER BY RANDOM()")
     List<Game> findAllRandomized();
 }
